修改价格小数点

+++ cofactory-1.1/Model/LookoverMaterialModel.h
@@ -17,7 +17,7 @@
 @property (nonatomic, copy) NSString *type;
 @property (nonatomic, assign) NSInteger userID;
 @property (nonatomic, assign) NSInteger materialID;
-@property (nonatomic, assign) NSInteger price;
+@property (nonatomic, assign) CGFloat price;
 @property (nonatomic, copy) NSString *useage;
 @property (nonatomic, copy) NSString *width;
 @property (nonatomic, copy) NSString *userName;
